made canvas a component and it now spans the width and height of the screen
